fix: row-major (buffer) vs column-major (host)
Because R uses column-major format we need to handle this on both ends of the conversion.
When sending the data from the buffer to the host we can specify the `byte_strides` to do this.

There does not seem to be an option to specify this when sending it back, as `host_layout` field cannot be used, at least not with the CPU PJRT plugin, when sending the data back to the host.

Therefore, we need to transpose the data manually.

I have created an associated issue in xla: https://github.com/openxla/xla/issues/28833