build: pjrt can now be built with -Wall -pedantic (#39)
Without this patch, we were getting errors when installing pjrt with pak::pak() which uses -Wall -pedantic, which is also used on CRAN.

This is the type of error:
  ../inst/include/xla/pjrt/c/pjrt_c_api.h:2506:26: error: declaration of ‘PJRT_Error* (* PJRT_Api::PJRT_Client_CreateUninitializedBuffer)(PJRT_Client_CreateUninitializedBuffer_Args*)’ changes meaning of ‘PJRT_Client_CreateUninitializedBuffer’ [-Wchanges-meaning]
   2506 |   _PJRT_API_STRUCT_FIELD(PJRT_Client_CreateUninitializedBuffer);

It happened because there was a typedef and a struct field with the same name, which was ok in C, but not in C++.
While we can disable the compiler flags that trigger this warning, this leads to rcmdcheck failure.

Patching the names of the API struct works, because C/++ does not reorder fields.