refactor: S7 instead of S3 generics, dtype -> elt_type
It's better to retrieve the `PJRTElementType` with `elt_type` than with `dtype`, because `dtype()` will also return `stablehlo::TensorDataType` in stablehlo, where it will also return a `TensorDataType` for `dtype<AnvilTensor>`, which is almost a `PJRTBuffer`. It would be confusing if two different types were returned for the same generic for almost the same object.